tests/test_authentication.py
"""Authentication tests for Huckleberry API."""

import asyncio
import time
from types import SimpleNamespace

import aiohttp
import pytest

from huckleberry_api import HuckleberryAPI


class TestAuthentication:
    """Test authentication functionality."""

    async def test_authenticate_invalid_credentials_includes_firebase_error_details(
        self, websession: aiohttp.ClientSession, monkeypatch: pytest.MonkeyPatch
    ) -> None:
        """Test authentication errors include Firebase response details."""

        class FakeResponse:
            def __init__(self) -> None:
                self.status = 400
                self.reason = "Bad Request"
                self.headers = {}
                self.request_info = SimpleNamespace(real_url="https://identitytoolkit.googleapis.com/")
                self.history = ()

            async def __aenter__(self) -> FakeResponse:
                return self

            async def __aexit__(self, exc_type, exc, tb) -> bool:
                return False

            async def text(self) -> str:
                return '{"error":{"message":"INVALID_LOGIN_CREDENTIALS","errors":[{"message":"INVALID_LOGIN_CREDENTIALS"}]}}'

            async def json(self, *args, **kwargs) -> dict[str, object]:
                return {
                    "error": {
                        "message": "INVALID_LOGIN_CREDENTIALS",
                        "errors": [{"message": "INVALID_LOGIN_CREDENTIALS"}],
                    }
                }

        invalid_api = HuckleberryAPI(
            email="invalid@test.com", password="wrongpassword", timezone="UTC", websession=websession
        )

        def fake_post(*args, **kwargs) -> FakeResponse:
            return FakeResponse()

        monkeypatch.setattr(websession, "post", fake_post)

        with pytest.raises(aiohttp.ClientResponseError) as exc_info:
            await invalid_api.authenticate()

        assert exc_info.value.status == 400
        assert "Authentication failed with HTTP 400 Bad Request" in exc_info.value.message
        assert (
            '{"error":{"message":"INVALID_LOGIN_CREDENTIALS","errors":[{"message":"INVALID_LOGIN_CREDENTIALS"}]}}'
            in exc_info.value.message
        )

    async def test_authenticate_success(self, api: HuckleberryAPI) -> None:
        """Test successful authentication."""
        assert api.id_token is not None
        assert api.refresh_token is not None
        assert api.user_uid is not None
        assert api.token_expires_at is not None

    async def test_authenticate_invalid_credentials(self, websession: aiohttp.ClientSession) -> None:
        """Test authentication with invalid credentials."""

        invalid_api = HuckleberryAPI(
            email="invalid@test.com", password="wrongpassword", timezone="UTC", websession=websession
        )
        with pytest.raises((RuntimeError, aiohttp.ClientResponseError, aiohttp.ClientError)):
            await invalid_api.authenticate()

    async def test_token_refresh(self, api: HuckleberryAPI) -> None:
        """Test token refresh functionality."""
        # Wait 1 second to ensure we get a new token (Firebase may return same token if too fresh)
        await asyncio.sleep(1)
        original_token = api.id_token
        await api.refresh_session_token()
        assert api.id_token is not None
        assert api.id_token != original_token

    async def test_ensure_session(self, api: HuckleberryAPI) -> None:
        """Test ensure_session keeps token validity."""
        original_token = api.id_token
        original_expires = api.token_expires_at

        # Call ensure_session - shouldn't refresh if token is fresh
        await api.ensure_session()
        assert api.id_token == original_token
        assert api.token_expires_at == original_expires

        # Simulate expired token by setting expiry to past
        old_expiry = time.time() - 100
        api.token_expires_at = old_expiry

        # Now ensure_session should refresh
        await api.ensure_session()
        assert api.id_token is not None
        assert api.token_expires_at is not None

        # Verify token was actually refreshed by checking expiry was updated significantly
        # New expiry should be at least 3000 seconds in the future (Firebase tokens are ~1 hour)
        assert api.token_expires_at > time.time() + 3000, (
            f"Token expiry not properly refreshed: was {old_expiry}, now {api.token_expires_at}"
        )

        # Verify the refreshed token works by making a Firestore call
        user_doc = await api.get_user()
        assert user_doc is not None


class TestChildrenRetrieval:
    """Test children data retrieval."""

    async def test_get_child(self, api: HuckleberryAPI) -> None:
        """Test retrieving user childList and child document by id."""
        user_doc = await api.get_user()
        assert user_doc is not None

        child_uids = [child_ref.cid for child_ref in user_doc.childList]
        assert len(child_uids) > 0
        assert len(child_uids) == len(set(child_uids))

        child = await api.get_child(child_uids[0])
        assert child is not None
        assert child.model_dump(exclude_none=True)


class TestErrorHandling:
    """Test error handling."""

    async def test_operations_require_authentication(self, websession: aiohttp.ClientSession) -> None:
        """Test that operations fail without authentication."""
        import os

        email = os.getenv("HUCKLEBERRY_EMAIL", "test@example.com")
        password = os.getenv("HUCKLEBERRY_PASSWORD", "password")
        timezone = os.getenv("HUCKLEBERRY_TIMEZONE", "UTC")

        unauthenticated_api = HuckleberryAPI(email=email, password=password, timezone=timezone, websession=websession)

        # Note: API actually requires authentication but doesn't always raise
        # Firestore SDK may succeed with cached credentials from fixture
        # This test verifies the API can be instantiated without immediate auth
        assert unauthenticated_api.id_token is None

    async def test_invalid_child_uid(self, api: HuckleberryAPI) -> None:
        """Test operations with invalid child UID."""
        # Firebase Security Rules block writes to invalid child UIDs
        # This is expected - it will raise PermissionDenied
        from google.api_core.exceptions import PermissionDenied

        with pytest.raises(PermissionDenied):
            await api.start_sleep("invalid-uid-12345")
